Update server.js
I don't really get all the functions (well I do a bit), but I (as bjkistad requested), have added a regular expression to ensure only alphanumeric characters are used in the username. This seems to be a strange attempt to prevent XSS, but it seems alright if we want to ensure control chars are not used, but in which case this should be done for all fields, including emails and passwords. I would like to point out that the error message file for having an invalid username has not been created yet, and should be at a later date. Scratch account: jokebookservice1