A Share Button Icon
It is blue (the colour of Newawe), and contains the text "Newawe" and "Share With" in big, and our motto in small. It can be used outside of Newawe, if people want to upload from other websites. Tell me whatyou think, and if it is good enough, maybe we could merge?